agente-ia/requirements.txt
flask==2.3.3
flask-cors==4.0.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
reportlab==4.0.7
pydantic==2.5.0
mcp==1.26.0